إضافة الفواصل في الكميات

+++ lib/features/control_panel/business/sections/invoice_actions.dart
@@ -180,7 +180,7 @@ class InvoiceActions {
   clearInvoiceDetailsIds() => _invoiceDetailsIds.clear();
 
   late WriteBatch _batch;
-  late int _remainedQuantity, _boughtQuantity;
+  late double _remainedQuantity, _boughtQuantity;
   late CollectionReference _invoiceColRef;
   late CollectionReference _productColRef;
